backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
channels>=4.0
channels-redis>=4.0  # optional for production
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
psycopg2-binary>=2.9 ; python_version >= '3.8'  # optional
python-decouple>=3.8
